Make Renovate re-create new suggestion PR on close
I am testing semantic versioning with Renovate for Terraform - Renovate
has already created a PR to suggest I update my dummy module to the
latest version, which is a major update. I will close that PR and re-run
Renovate, it should not create a PR for the major update to my module,
only one PR for the minor update that is available (v0.1.0 -> v0.1.1).